backend/requirements.txt
Django==5.2.3
djangorestframework==3.15.2
django-cors-headers==4.6.0
psycopg2-binary==2.9.9
python-decouple==3.8
djangorestframework-simplejwt==5.3.0
django-filter==24.3
google-generativeai==0.8.3
python-dotenv==1.0.1
Pillow==10.4.0
